# docs/api/updated-openapi.json
{"openapi":"3.1.0","info":{"title":"Saleor Price Manager","description":"FastAPI service for managing dynamic pricing across multiple Saleor channels with percentage-based markups and Redis caching","version":"1.0.0"},"paths":{"/api/channels/":{"get":{"tags":["channels"],"summary":"List All Channels","description":"Retrieve a list of all Saleor channels with their current markup percentages.\n    \n    This endpoint queries Saleor API for channel information and enriches it with\n    markup data from Redis cache or Saleor metadata.\n    \n    **Returns:**\n    - List of channels with markup information\n    - Each channel includes: id, name, slug, markup_percent, metadata\n    \n    **Authentication:** Bearer token required","operationId":"list_channels_api_channels__get","responses":{"200":{"description":"Successfully retrieved channels list","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ChannelWithMarkup"},"type":"array","title":"Response List Channels Api Channels  Get"}}}},"401":{"description":"Authentication required or token invalid"},"500":{"description":"Internal server error"}},"security":[{"HTTPBearer":[]}]}},"/api/channels/markup":{"post":{"tags":["channels"],"summary":"Set Channel Markup","description":"Set or update the percentage markup for a specific channel.\n    \n    This operation will:\n    1. Update the channel metadata in Saleor via GraphQL API\n    2. Update the Redis cache for immediate access\n    3. Trigger price recalculation for products in this channel\n    \n    **Parameters:**\n    - channel_id: Base64 encoded Saleor channel ID\n    - markup_percent: Percentage markup to apply (0-1000)\n    \n    **Authentication:** Bearer token required","operationId":"set_markup_api_channels_markup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelMarkup"}}},"required":true},"responses":{"200":{"description":"Markup successfully updated","content":{"application/json":{"schema":{}}}},"400":{"description":"Failed to update channel markup"},"401":{"description":"Authentication required or token invalid"},"422":{"description":"Request validation failed"}},"security":[{"HTTPBearer":[]}]}},"/api/prices/calculate":{"post":{"tags":["prices"],"summary":"Calculate Product Price","description":"Calculate the final price for a product in a specific channel,\n    applying the channel's markup percentage to the base price.\n    \n    **Formula:** `final_price = base_price * (1 + markup_percent / 100)`\n    \n    **Parameters:**\n    - product_id: Base64 encoded Saleor product ID\n    - channel_id: Base64 encoded Saleor channel ID  \n    - base_price: Original price before markup\n    \n    **Returns:**\n    - Detailed price calculation including base price, markup, and final price\n    \n    **Authentication:** Bearer token required","operationId":"calculate_price_api_prices_calculate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PriceCalculationRequest"}}},"required":true},"responses":{"200":{"description":"Price calculated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PriceCalculationResponse"}}}},"400":{"description":"Calculation error"},"401":{"description":"Authentication required or token invalid"},"422":{"description":"Request validation failed"}},"security":[{"HTTPBearer":[]}]}},"/api/prices/batch-calculate":{"post":{"tags":["prices"],"summary":"Batch Calculate Prices","description":"Calculate prices for multiple products across different channels in a single request.\n    \n    Useful for bulk price updates or when processing large product catalogs.\n    \n    **Parameters:**\n    - Array of price calculation requests, each containing:\n      - product_id: Base64 encoded Saleor product ID\n      - channel_id: Base64 encoded Saleor channel ID\n      - base_price: Original price before markup\n    \n    **Returns:**\n    - Array of detailed price calculations\n    \n    **Authentication:** Bearer token required","operationId":"batch_calculate_api_prices_batch_calculate_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/PriceCalculationRequest"},"type":"array","title":"Items"}}},"required":true},"responses":{"200":{"description":"Batch calculation completed","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/PriceCalculationResponse"},"type":"array","title":"Response Batch Calculate Api Prices Batch Calculate Post"}}}},"400":{"description":"Batch calculation error"},"401":{"description":"Authentication required or token invalid"},"422":{"description":"Request validation failed"}},"security":[{"HTTPBearer":[]}]}},"/webhooks/product-updated":{"post":{"tags":["webhooks"],"summary":"Handle Product Updated Webhook","description":"Webhook endpoint for Saleor PRODUCT_UPDATED events.\n    \n    When a product is updated in Saleor, this endpoint triggers\n    background price recalculation for all channels.\n    \n    **Event Processing:**\n    1. Validates the webhook payload\n    2. Queues background task for price recalculation\n    3. Updates prices across all channels for the product\n    \n    **Authentication:** No bearer token required (webhook signatures handled separately)\n    \n    **Background Processing:** Price calculations are performed asynchronously","operationId":"handle_product_updated_webhooks_product_updated_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaleorWebhookPayload"}}},"required":true},"responses":{"200":{"description":"Webhook received and processed","content":{"application/json":{"schema":{},"example":{"status":"received"}}}},"400":{"description":"Invalid webhook payload","content":{"application/json":{"example":{"detail":"Invalid webhook payload"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/webhooks/channel-created":{"post":{"tags":["webhooks"],"summary":"Handle Channel Created Webhook","description":"Webhook endpoint for Saleor CHANNEL_CREATED events.\n    \n    When a new channel is created in Saleor, this endpoint\n    invalidates related caches and prepares the channel for markup management.\n    \n    **Event Processing:**\n    1. Validates the webhook payload\n    2. Invalidates Redis cache for the new channel\n    3. Prepares channel for markup configuration\n    \n    **Authentication:** No bearer token required (webhook signatures handled separately)","operationId":"handle_channel_created_webhooks_channel_created_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaleorWebhookPayload"}}},"required":true},"responses":{"200":{"description":"Webhook received and processed","content":{"application/json":{"schema":{},"example":{"status":"received"}}}},"400":{"description":"Invalid webhook payload","content":{"application/json":{"example":{"detail":"Invalid webhook payload"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ChannelMarkup":{"properties":{"channel_id":{"type":"string","title":"Channel Id","description":"Base64 encoded Saleor channel ID","example":"Q2hhbm5lbDox"},"markup_percent":{"anyOf":[{"type":"number","maximum":1000.0,"minimum":0.0},{"type":"string"}],"title":"Markup Percent","description":"Markup percentage to apply (0-1000)","example":15.5}},"type":"object","required":["channel_id","markup_percent"],"title":"ChannelMarkup","description":"Request model for setting channel markup percentage"},"ChannelWithMarkup":{"properties":{"id":{"type":"string","title":"Id","description":"Base64 encoded Saleor channel ID"},"name":{"type":"string","title":"Name","description":"Human-readable channel name"},"slug":{"type":"string","title":"Slug","description":"URL-friendly channel identifier"},"markup_percent":{"type":"string","title":"Markup Percent","description":"Current markup percentage"},"metadata":{"anyOf":[{"items":{"additionalProperties":{"type":"string"},"type":"object"},"type":"array"},{"type":"null"}],"title":"Metadata","description":"Channel metadata from Saleor"}},"type":"object","required":["id","name","slug","markup_percent"],"title":"ChannelWithMarkup","description":"Channel information with markup data"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PriceCalculationRequest":{"properties":{"product_id":{"type":"string","title":"Product Id","description":"Base64 encoded Saleor product ID","example":"UHJvZHVjdDox"},"channel_id":{"type":"string","title":"Channel Id","description":"Base64 encoded Saleor channel ID","example":"Q2hhbm5lbDox"},"base_price":{"anyOf":[{"type":"number"},{"type":"string"}],"title":"Base Price","description":"Base price before markup","example":100.0}},"type":"object","required":["product_id","channel_id","base_price"],"title":"PriceCalculationRequest","description":"Request model for price calculation"},"PriceCalculationResponse":{"properties":{"product_id":{"type":"string","title":"Product Id","description":"Base64 encoded Saleor product ID"},"channel_id":{"type":"string","title":"Channel Id","description":"Base64 encoded Saleor channel ID"},"base_price":{"type":"string","title":"Base Price","description":"Original base price"},"markup_percent":{"type":"string","title":"Markup Percent","description":"Applied markup percentage"},"final_price":{"type":"string","title":"Final Price","description":"Final price after markup"},"currency":{"type":"string","title":"Currency","description":"Currency code","default":"USD"}},"type":"object","required":["product_id","channel_id","base_price","markup_percent","final_price"],"title":"PriceCalculationResponse","description":"Response model for price calculation"},"SaleorWebhookPayload":{"properties":{"event_type":{"type":"string","title":"Event Type","description":"Type of Saleor event","example":"PRODUCT_UPDATED"},"product_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Product Id","description":"Product ID for product-related events","example":"UHJvZHVjdDox"},"channel_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Channel Id","description":"Channel ID for channel-related events","example":"Q2hhbm5lbDox"},"data":{"additionalProperties":true,"type":"object","title":"Data","description":"Additional event data from Saleor"}},"type":"object","required":["event_type"],"title":"SaleorWebhookPayload","description":"Saleor webhook event payload"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}